Strimlite_Version_V1/requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
sentence-transformers
faiss-cpu
python-dotenv
torch